feat: 智能移调优化 - 时值力度加权 + 全半音移调
- 添加 MIDI velocity 解析和存储
- 直方图使用 duration * velocity 加权
- 移调算法改为尝试所有半音移调（-48~+48）
- 不再局限于八度移调，更精确地选择最佳位置

+++ src/core/PlaybackEngine.h
@@ -108,7 +108,8 @@ namespace Core {
         std::unordered_map<std::pair<void*, int>, TempNote*, PairHash> m_active_notes_map;
         
         // Optimization: Cache pitch statistics per track (index = track_idx)
-        std::vector<std::vector<int>> m_track_pitch_histograms;
+        // Using float for duration-weighted histogram
+        std::vector<std::vector<float>> m_track_pitch_histograms;
         
         std::atomic<int> m_config_version{0};   // To trigger rebuild
         int m_built_version{-1};                // Last built version
